requirements.txt
aiohttp~=3.10.10
asyncpg~=0.30.0
pydantic~=2.9.2
pytest
pytest-aiohttp
pytest-mock
pyjwt~=2.9.0
python-dotenv~=1.0.1
flasgger